fix: rename footer.tsx to index.tsx (#1926)
