+++ core/components/footer/index.tsx
